requirements.txt
# Core Azure SDKs
azure-communication-email==1.1.0
azure-core==1.36.0
azure-cosmos==4.14.0
azure-functions==1.24.0
azure-identity==1.25.1
azure-servicebus==7.14.2

# Flask app
Flask==3.1.2
Bootstrap-Flask==2.5.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.2
WTForms==3.2.1

# Utilities
python-dotenv==1.1.1
requests==2.32.5

# Enforce Python runtime
python_version >= "3.12"
